--- src/main/java/com/prgrms/bdbks/global/util/UrlValidator.java
@@ -1,15 +1,0 @@
-package com.prgrms.bdbks.global.util;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-public class UrlValidator {
-
-    private static org.apache.commons.validator.routines.UrlValidator validator = new org.apache.commons.validator.routines.UrlValidator();
-
-    public static boolean isValid(String url) {
-        return validator.isValid(url);
-    }
-
-}
